+++ app/Plugin/Admin/Directory_Listing.php
@@ -371,23 +371,23 @@ class Directory_Listing {
 					'label' => __( 'Show directory', 'external-files-in-media-library' ),
 				),
 			),
-			'aws_s3_api'            => array(
+			'aws_s3_api'                    => array(
 				'title'            => __( 'Enter your credentials', 'external-files-in-media-library' ),
 				'description'      => __( 'Use the login details for your IAM user who has permissions for the bucket you are using. See:', 'external-files-in-media-library' ) . ' <a href="https://docs.aws.amazon.com/AmazonS3/latest/userguide/security-iam.html" target="_blank">https://docs.aws.amazon.com/AmazonS3/latest/userguide/security-iam.html</a>',
-				'access_key'              => array(
-					'label' => __( 'Access Key', 'external-files-in-media-library' )
+				'access_key'       => array(
+					'label' => __( 'Access Key', 'external-files-in-media-library' ),
 				),
-				'secret_key'              => array(
-					'label' => __( 'Secret Key', 'external-files-in-media-library' )
+				'secret_key'       => array(
+					'label' => __( 'Secret Key', 'external-files-in-media-library' ),
 				),
-				'bucket'              => array(
-					'label' => __( 'Bucket', 'external-files-in-media-library' )
+				'bucket'           => array(
+					'label' => __( 'Bucket', 'external-files-in-media-library' ),
 				),
 				'save_credentials' => array(
-					'label' => __( 'Save this credentials in directory archive', 'external-files-in-media-library' )
+					'label' => __( 'Save this credentials in directory archive', 'external-files-in-media-library' ),
 				),
 				'button'           => array(
-					'label' => __( 'Show directory', 'external-files-in-media-library' )
+					'label' => __( 'Show directory', 'external-files-in-media-library' ),
 				),
 			),
 			'services'                      => array(